+++ src/KTrie/KTrie.csproj
@@ -9,6 +9,7 @@
     <PackageProjectUrl>https://github.com/kpol/trie</PackageProjectUrl>
     <PackageTags>trie, prefix-tree, data-structure</PackageTags>
     <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
 </Project>
